AP-17454: Fixed parsing of truncated (hence invalid) XML for big XML cells in XPath node dialog
Fixed the problem with the XPath node when a XML File has got more than
2000 lines which made it fail due to changes which came with Java11.
Catching now the SAXException instead of StopParseException which was
only thrown when having a english locale but i.e. not when having a
german locale. In the german locale case a SAXParseException will be
thrown for whatever reason.

AP-17454 (XML Path node fails to parse XML documents with more than 2000 lines on non-English locale)